Repositioned pageHeader into a dedicated component. Integrate ability to share pages in social media.
